app/requirements.txt
dataclasses
ConfigParser
paho-mqtt
requests
types-requests
python-dotenv
schedule
